gateway/blueprints/requirements.txt
requests
prometheus_client
flask
jinja2
gunicorn
gevent
Pillow
numpy
google
protobuf
grpcio
grpcio-tools
pyyaml